Refactor widgets to separate files
